% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{Ackley}
\alias{Ackley}
\title{Ackley benchmark function
f(x) = 0 at x = (0, ..., 0)
x_i in [-32.768, 32.768]}
\usage{
Ackley(x)
}
\arguments{
\item{x}{\itemize{
\item numeric
}}
}
\value{
numeric value of Ackley function
}
\description{
Ackley benchmark function
f(x) = 0 at x = (0, ..., 0)
x_i in [-32.768, 32.768]
}
