% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{Baele}
\alias{Baele}
\title{Baele
Min: f(x) = 0 at x = (3, 0.5)
x1, x2 in [-4.5, 4.5]}
\usage{
Baele(x)
}
\arguments{
\item{x}{\itemize{
\item numeric
}}
}
\value{

}
\description{
Baele
Min: f(x) = 0 at x = (3, 0.5)
x1, x2 in [-4.5, 4.5]
}
