% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ela_classification.R
\name{calculate_ela_features}
\alias{calculate_ela_features}
\title{Calculate selected ELA features for given fitness.}
\usage{
calculate_ela_features(fitness, lower, upper)
}
\arguments{
\item{fitness}{fitness function, that returns a numerical value.}

\item{lower}{numeric - lower bound of the domain, a vector of length equal
to the decision variables.}

\item{upper}{numeric - upper bound of the domain, a vector of length equal
to the decision variables.}
}
\value{
Returns a named list with values of selected ELA features.
}
\description{
Calculate selected ELA features for given fitness.
}
