% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ela_classification.R
\name{classify_optimization_problem}
\alias{classify_optimization_problem}
\title{Classify optimization problem using selected ELA features and Random Forest model trained on BBOB dataset.}
\usage{
classify_optimization_problem(fitness, lower, upper)
}
\arguments{
\item{fitness}{fitness function, that returns a numerical value, to be classified. The domain should be at least two dimensional.}

\item{lower}{numeric - lower bound of the domain, a vector of length equal
to the decision variables.}

\item{upper}{numeric - upper bound of the domain, a vector of length equal
to the decision variables.}
}
\value{
Returns one of c("low-conditioning", "multimodal-adequate", "multimodal-weak", "separable", "unimodal").
}
\description{
Classify optimization problem using selected ELA features and Random Forest model trained on BBOB dataset.
}
\examples{
f <- function(x) x[[1]] + x[[2]]
result <- classify_optimization_problem(fitness = f, lower = c(-5, -5), upper = c(5, 5))
}
