% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_method.R
\name{default_run_gradient_method}
\alias{default_run_gradient_method}
\title{Function that runs gradient method for one deme.}
\usage{
default_run_gradient_method(deme, fitness, optim_args)
}
\arguments{
\item{deme}{\itemize{
\item Deme
}}

\item{fitness}{\itemize{
\item function
}}

\item{optim_args}{\itemize{
\item list of additional parameters
}}
}
\description{
Function that runs gradient method for one deme.
}
