% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deoptim_metaepoch.R
\name{deoptim_metaepoch}
\alias{deoptim_metaepoch}
\title{Function that runs one differential evolution metaepoch. Wrapper function for DEoptim::DEoptim.}
\usage{
deoptim_metaepoch(config_deoptim)
}
\arguments{
\item{config_deoptim}{\itemize{
\item list of DEoptim::DEoptim params
}}
}
\value{
list with named fields: solution, population, value. See
\code{\link{ga_metaepoch}} for more details.
}
\description{
Function that runs one differential evolution metaepoch. Wrapper function for DEoptim::DEoptim.
}
\examples{
tree_height <- 3
empty_config_deoptim <- lapply(1:tree_height, function(x) {
  list()
})
deoptim_metaepoch(empty_config_deoptim)
}
