% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\name{Deprecated}
\alias{Deprecated}
\alias{is.hms}
\alias{as.hms}
\alias{as.hms.default}
\alias{as.hms.POSIXt}
\alias{as.hms.POSIXlt}
\title{Deprecated functions}
\usage{
is.hms(x)

as.hms(x, ...)

\method{as.hms}{default}(x, ...)

\method{as.hms}{POSIXt}(x, tz = pkgconfig::get_config("hms::default_tz", ""), ...)

\method{as.hms}{POSIXlt}(x, tz = pkgconfig::get_config("hms::default_tz", ""), ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Arguments passed on to further methods.}

\item{tz}{The time zone in which to interpret a POSIXt time for extracting
the time of day.  The default is now the zone of \code{x} but was \code{"UTC"}
for v0.3 and earlier.  The previous behavior can be restored by calling
\code{pkgconfig::set_config("hms::default_tz", "UTC")}, see
\code{\link[pkgconfig:set_config]{pkgconfig::set_config()}}.}
}
\description{
\code{is.hms()} has been replaced by \code{\link[=is_hms]{is_hms()}}.

\code{as.hms()} has been replaced by \code{\link[=as_hms]{as_hms()}}, which is no longer generic and also
does not have a \code{tz} argument.
It also uses the time zone of the argument for conversion,
not the current system's timezone.
Change the timezone before converting if necessary, e.g. using \code{\link[lubridate:with_tz]{lubridate::with_tz()}}.
}
\keyword{internal}
