% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_metaepoch.R
\name{ecr_metaepoch}
\alias{ecr_metaepoch}
\title{Function that runs one ecr metaepoch.}
\usage{
ecr_metaepoch(config_ecr)
}
\arguments{
\item{config_ecr}{\itemize{
\item list of ecr::ecr params
}}
}
\value{
list with named fields: solution, population, value
}
\description{
Function that runs one ecr metaepoch.
}
