% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{Eggholder}
\alias{Eggholder}
\title{Eggholder
Min: f(x) = -959.6407 at x = (512, 404.2319)
x1, x2 in [-512, 512]}
\usage{
Eggholder(x)
}
\arguments{
\item{x}{\itemize{
\item numeric
}}
}
\value{

}
\description{
Eggholder
Min: f(x) = -959.6407 at x = (512, 404.2319)
x1, x2 in [-512, 512]
}
