% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ela_features}
\alias{ela_features}
\title{Exploratory Landscape Analysis (ELA) features dataset}
\format{
A data frame with 1920 rows and 5 columns:
\describe{
\item{ic.eps.ratio}{ratio of partial information sensitivity}
\item{nbc.nb_fitness.cor}{correlation between fitness value and count of observations to whom the current observation is the nearest better neighbour}
\item{ela_meta.quad_simple.adj_r2}{adjusted R^2 of a quadratic model without interactions}
\item{ela_meta.lin_simple.adj_r2}{adjusted R^2 of a simple linear model}
\item{type}{One of c("low-conditioning", "multimodal-adequate", "multimodal-weak", "separable", "unimodal")}
}
}
\usage{
ela_features
}
\description{
The ela_features dataset, created using the flacco package, includes ELA
features for all BBOB functions (all available fids) across different
dimensions (2, 5, 10, 20) and instances (1:20).
Key features captured are ic.eps.ratio, nbc.nb_fitness.cor,
ela_meta.quad_simple.adj_r2, and ela_meta.lin_simple.adj_r2, along with
function type. This dataset is essential for training classification models.
In case of hmsr package, the dataset is used to train Random Forest model.
}
\keyword{datasets}
