% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_cma_es_metaepoch.R
\name{ga_cma_es_metaepoch}
\alias{ga_cma_es_metaepoch}
\title{Function that generates run_metaepoch function for two level HMS.
First level: GA, second level: CMA-ES.}
\usage{
ga_cma_es_metaepoch(ga_cma_es_config)
}
\arguments{
\item{ga_cma_es_config}{\itemize{
\item list that consists of two lists: GA params and CMA-ES params.
}}
}
\value{
list with named fields: solution, population, value. See
\code{\link{ga_metaepoch}} for more details.
}
\description{
Function that generates run_metaepoch function for two level HMS.
First level: GA, second level: CMA-ES.
}
\examples{
tree_height <- 2
ga_config <- list()
cma_es_config <- list()
config <- list(ga_config, cma_es_config)
ga_cma_es_metaepoch(config)
}
