% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_stopping_condition.R
\name{global_stopping_condition_max_fitness_evaluations}
\alias{global_stopping_condition_max_fitness_evaluations}
\title{global_stopping_condition_max_fitness_evaluations}
\usage{
global_stopping_condition_max_fitness_evaluations(max_evaluations)
}
\arguments{
\item{max_evaluations}{\itemize{
\item numeric
}}
}
\value{

}
\description{
global_stopping_condition_max_fitness_evaluations
}
