% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_stopping_condition.R
\name{global_stopping_condition_metaepochs_count}
\alias{global_stopping_condition_metaepochs_count}
\title{global_stopping_condition_metaepochs_count}
\usage{
global_stopping_condition_metaepochs_count(metaepochs_count)
}
\arguments{
\item{metaepochs_count}{\itemize{
\item numeric
}}
}
\value{

}
\description{
global_stopping_condition_metaepochs_count
}
