% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_stopping_condition.R
\name{global_stopping_condition_trivial}
\alias{global_stopping_condition_trivial}
\title{global_stopping_condition_trivial}
\usage{
global_stopping_condition_trivial()
}
\value{

}
\description{
global_stopping_condition_trivial
}
