% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{Griewank}
\alias{Griewank}
\title{Griewank
Min: f(x) = 0 at x = (0, ..., 0)
x_i in [-600, 600]}
\usage{
Griewank(x)
}
\arguments{
\item{x}{\itemize{
\item numeric
}}
}
\value{
numeric value of Griewank function
}
\description{
Griewank
Min: f(x) = 0 at x = (0, ..., 0)
x_i in [-600, 600]
}
