% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_stopping_condition.R
\name{gsc_max_fitness_evaluations}
\alias{gsc_max_fitness_evaluations}
\title{gsc_max_fitness_evaluations}
\usage{
gsc_max_fitness_evaluations(max_evaluations)
}
\arguments{
\item{max_evaluations}{\itemize{
\item numeric - maximum number of fitness function evaluations
}}
}
\value{
function that can be used as a global stopping condition
}
\description{
gsc_max_fitness_evaluations
}
