% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_stopping_condition.R
\name{gsc_metaepochs_count}
\alias{gsc_metaepochs_count}
\title{gsc_metaepochs_count}
\usage{
gsc_metaepochs_count(metaepochs_count)
}
\arguments{
\item{metaepochs_count}{\itemize{
\item numeric
}}
}
\value{

}
\description{
gsc_metaepochs_count
}
