% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\docType{class}
\name{hms-class}
\alias{hms-class}
\title{hms}
\description{
hms
}
\section{Slots}{

\describe{
\item{\code{root_id}}{character.}

\item{\code{metaepoch_snapshots}}{list.}

\item{\code{best_fitness}}{numeric.}

\item{\code{best_solution}}{numeric.}

\item{\code{total_time_in_seconds}}{numeric.}

\item{\code{total_metaepoch_time_in_seconds}}{numeric.}

\item{\code{metaepochs_count}}{numeric.}

\item{\code{deme_population_sizes}}{numeric.}

\item{\code{lower}}{numeric.}

\item{\code{upper}}{numeric.}

\item{\code{call}}{language.}
}}

