% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stopping_condition.R
\name{local_stopping_condition_max_fitness_evaluations}
\alias{local_stopping_condition_max_fitness_evaluations}
\title{local_stopping_condition_max_fitness_evaluations}
\usage{
local_stopping_condition_max_fitness_evaluations(max_evaluations_count)
}
\arguments{
\item{max_evaluations_count}{\itemize{
\item numeric
}}
}
\value{

}
\description{
local_stopping_condition_max_fitness_evaluations
}
