% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stopping_condition.R
\name{local_stopping_condition_metaepochs_without_improvement}
\alias{local_stopping_condition_metaepochs_without_improvement}
\title{local_stopping_condition_metaepochs_without_improvement}
\usage{
local_stopping_condition_metaepochs_without_improvement(
  max_metaepochs_without_improvement
)
}
\arguments{
\item{max_metaepochs_without_improvement}{\itemize{
\item numeric
}}
}
\value{

}
\description{
local_stopping_condition_metaepochs_without_improvement
}
