% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stopping_condition.R
\name{local_stopping_condition_trivial}
\alias{local_stopping_condition_trivial}
\title{local_stopping_condition_trivial}
\usage{
local_stopping_condition_trivial()
}
\value{

}
\description{
local_stopping_condition_trivial
}
