% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stopping_condition.R
\name{lsc_max_fitness_evaluations}
\alias{lsc_max_fitness_evaluations}
\title{lsc_max_fitness_evaluations}
\usage{
lsc_max_fitness_evaluations(max_evaluations)
}
\arguments{
\item{max_evaluations}{\itemize{
\item numeric
}}
}
\value{
function that can be used as a local stopping condition
}
\description{
lsc_max_fitness_evaluations
}
