% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stopping_condition.R
\name{lsc_metaepochs_without_active_child}
\alias{lsc_metaepochs_without_active_child}
\title{Factory function for a local stopping condition that stops a deme
after given number of metaepochs have past since last metaepoch during
which this deme had an active child.}
\usage{
lsc_metaepochs_without_active_child(metaepochs_limit)
}
\arguments{
\item{metaepochs_limit}{\itemize{
\item number of metaepochs that a deme can be active
without any active child
}}
}
\value{
Function that can be used as a local stopping condition for hms.
}
\description{
Factory function for a local stopping condition that stops a deme
after given number of metaepochs have past since last metaepoch during
which this deme had an active child.
}
\examples{
local_stopping_condition <- lsc_metaepochs_without_active_child(3)
}
