% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stopping_condition.R
\name{lsc_metaepochs_without_improvement}
\alias{lsc_metaepochs_without_improvement}
\title{lsc_metaepochs_without_improvement}
\usage{
lsc_metaepochs_without_improvement(max_metaepochs_without_improvement)
}
\arguments{
\item{max_metaepochs_without_improvement}{\itemize{
\item numeric
}}
}
\value{
function that can be used as a local stopping condition
}
\description{
lsc_metaepochs_without_improvement
}
