% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{max_euclidean_distance_sprouting_condition}
\alias{max_euclidean_distance_sprouting_condition}
\title{Default sprouting condition based on euclidean distance.}
\usage{
max_euclidean_distance_sprouting_condition(max_distance)
}
\arguments{
\item{max_distance}{\itemize{
\item numeric
}}
}
\value{
logical
}
\description{
Default sprouting condition based on euclidean distance.
}
