% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{max_metric_sprouting_condition}
\alias{max_metric_sprouting_condition}
\title{Default sprouting condition based on given metric.}
\usage{
max_metric_sprouting_condition(metric, max_distance)
}
\arguments{
\item{metric}{\itemize{
\item Metric used for deme distance comparison
}}

\item{max_distance}{\itemize{
\item numeric
}}
}
\value{
logical
}
\description{
Default sprouting condition based on given metric.
}
