% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\docType{class}
\name{MetaepochSnapshot-class}
\alias{MetaepochSnapshot-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{demes}}{list.}

\item{\code{best_fitness}}{numeric.}

\item{\code{best_solution}}{numeric.}

\item{\code{time_in_seconds}}{numeric.}

\item{\code{fitness_evaluations}}{numeric.}

\item{\code{blocked_sprouts}}{list.}

\item{\code{is_evolutionary}}{logical.}
}}

