% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_hms}
\alias{parse_hms}
\alias{parse_hm}
\title{Parsing hms values}
\usage{
parse_hms(x)

parse_hm(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
An object of class \link{hms}.
}
\description{
These functions convert character vectors to objects of the \link{hms} class.
\code{NA} values are supported.

\code{parse_hms()} accepts values of the form \code{"HH:MM:SS"}, with optional fractional seconds.

\code{parse_hm()} accepts values of the form \code{"HH:MM"}.
}
\examples{
parse_hms("12:34:56")
parse_hms("12:34:56.789")
parse_hm("12:34")
}
