% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{plot,hms-method}
\alias{plot,hms-method}
\title{Plot method for "hms" class.}
\usage{
\S4method{plot}{hms}(x)
}
\arguments{
\item{x}{\itemize{
\item hms s4 object
}}
}
\value{
It doesn't return anything meaningful. It plots the fitness by metaepoch count.
}
\description{
Plot method for "hms" class.
}
\examples{
f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
plot(result)
}
