% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{plotActiveDemes,hms-method}
\alias{plotActiveDemes,hms-method}
\title{plotActiveDemes}
\usage{
\S4method{plotActiveDemes}{hms}(object)
}
\arguments{
\item{object}{\itemize{
\item hms s4 object
}}
}
\description{
plotActiveDemes
}
