% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{plotPopulation}
\alias{plotPopulation}
\title{plotPopulation}
\usage{
plotPopulation(object, dimensions)
}
\arguments{
\item{object}{\itemize{
\item hms s4 object
}}

\item{dimensions}{\itemize{
\item two selected dimensions
}}
}
\description{
plotPopulation
}
