% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{print,hms-method}
\alias{print,hms-method}
\title{Print method for class "hms".}
\usage{
\S4method{print}{hms}(x, ...)
}
\arguments{
\item{x}{\itemize{
\item hms s4 object
}}

\item{...}{\itemize{
\item other print arguments
}}
}
\value{
It does not return anything. The obvious side effect is output to the terminal.
}
\description{
Print method for class "hms".
}
\examples{
f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
print(result)
}
