% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{printBlockedSprouts,hms-method}
\alias{printBlockedSprouts,hms-method}
\title{printBlockedSprouts}
\usage{
\S4method{printBlockedSprouts}{hms}(object)
}
\arguments{
\item{object}{\itemize{
\item hms s4 object
}}
}
\description{
printBlockedSprouts
}
