% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{printBlockedSprouts}
\alias{printBlockedSprouts}
\title{printBlockedSprouts}
\usage{
printBlockedSprouts(object)
}
\arguments{
\item{object}{\itemize{
\item hms s4 object
}}
}
\description{
printBlockedSprouts
}
