% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{printTree}
\alias{printTree}
\title{printTree method for class "hms".}
\usage{
printTree(object)
}
\arguments{
\item{object}{\itemize{
\item hms s4 object
}}
}
\value{
It does not return anything. It prints the hms tree.
}
\description{
printTree method for class "hms".
}
\examples{
f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
printTree(result)
}
