% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{Rastrigin}
\alias{Rastrigin}
\title{Rastrigin benchmark function
f(x) = 0 at x = (0, ..., 0)
x_i in  [-5.12, 5.12]}
\usage{
Rastrigin(x)
}
\arguments{
\item{x}{\itemize{
\item numeric
}}
}
\value{
numeric value of Rastrigin function
}
\description{
Rastrigin benchmark function
f(x) = 0 at x = (0, ..., 0)
x_i in  [-5.12, 5.12]
}
