% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgenoud_metaepoch.R
\name{rgenoud_metaepoch}
\alias{rgenoud_metaepoch}
\title{Function that runs one GA metaepoch.}
\usage{
rgenoud_metaepoch(config_rgenoud)
}
\arguments{
\item{config_rgenoud}{\itemize{
\item list of rgenoud::genoud params
}}
}
\value{
list with named fields: solution, population, value
}
\description{
Function that runs one GA metaepoch.
}
