% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{rtnorm_mutation}
\alias{rtnorm_mutation}
\title{rtnorm mutation - truncated normal distribution}
\usage{
rtnorm_mutation(lower, upper, sd)
}
\arguments{
\item{lower}{\itemize{
\item numeric
}}

\item{upper}{\itemize{
\item numeric
}}

\item{sd}{\itemize{
\item numeric
}}
}
\value{
function that can be used as a GA mutation function
}
\description{
rtnorm mutation - truncated normal distribution
}
