% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{saveMetaepochsPopulations}
\alias{saveMetaepochsPopulations}
\title{Title}
\usage{
saveMetaepochsPopulations(object, path, dimensions)
}
\arguments{
\item{object}{\itemize{
\item hms s4 object
}}

\item{path}{\itemize{
\item path
}}

\item{dimensions}{\itemize{
\item two selected dimensions
}}
}
\description{
Title
}
