% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprouting_condition.R
\name{sc_max_metric}
\alias{sc_max_metric}
\title{Default sprouting condition based on given metric.}
\usage{
sc_max_metric(metric, max_distances)
}
\arguments{
\item{metric}{\itemize{
\item Metric used for deme distance comparison
}}

\item{max_distances}{\itemize{
\item numeric
}}
}
\value{
logical
}
\description{
Default sprouting condition based on given metric.
}
