% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{Schwefel}
\alias{Schwefel}
\title{Schwefel benchmark function
Min: f(x) = 0 at x = (420.9687, ..., 420.9687)
x_i in [-500, 500]}
\usage{
Schwefel(x)
}
\arguments{
\item{x}{\itemize{
\item numeric
}}
}
\value{
numeric value of Schwefel function
}
\description{
Schwefel benchmark function
Min: f(x) = 0 at x = (420.9687, ..., 420.9687)
x_i in [-500, 500]
}
