% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{summary,hms-method}
\alias{summary,hms-method}
\title{Title}
\usage{
\S4method{summary}{hms}(object, ...)
}
\arguments{
\item{object}{\itemize{
\item hms s4 object
}}

\item{...}{\itemize{
\item other summary arguments
}}
}
\description{
Title
}
