% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ela_classification.R
\name{train_random_forest_model}
\alias{train_random_forest_model}
\title{Train Random Forest model on BBOB dataset (data/ela_features.rda).}
\usage{
train_random_forest_model()
}
\value{
Returns a Random Forest model trained on data/ela_features.rda dataset.
}
\description{
Train Random Forest model on BBOB dataset (data/ela_features.rda).
}
