% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{vec_cast.hms}
\alias{vec_cast.hms}
\title{Casting}
\usage{
\method{vec_cast}{hms}(x, to)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}
}
\description{
Double dispatch methods to support \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}}.
}
