% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{vec_type2.hms}
\alias{vec_type2.hms}
\title{Coercion}
\usage{
\method{vec_type2}{hms}(x, y, ..., x_arg = "", y_arg = "")
}
\arguments{
\item{x}{Either vector types; i.e.}

\item{y}{Either vector types; i.e.}
}
\description{
Double dispatch methods to support \code{\link[vctrs:vec_type2]{vctrs::vec_type2()}}.
}
