% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecr_metaepoch.R
\name{ecr_metaepoch}
\alias{ecr_metaepoch}
\title{Function that runs one ecr metaepoch. Wrapper function for ecr::ecr.}
\usage{
ecr_metaepoch(config_ecr)
}
\arguments{
\item{config_ecr}{\itemize{
\item list of ecr::ecr params
}}
}
\value{
list with named fields: solution, population, value. See
\code{\link{ga_metaepoch}} for more details.
}
\description{
Function that runs one ecr metaepoch. Wrapper function for ecr::ecr.
}
\examples{
tree_height <- 3
empty_config_ecr <- lapply(1:tree_height, function(x) {
  list()
})
ecr_metaepoch(empty_config_ecr)
}
