% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Euclidean distance}
\usage{
euclidean_distance(x, y)
}
\arguments{
\item{x}{\itemize{
\item numeric
}}

\item{y}{\itemize{
\item numeric
}}
}
\value{
numeric - euclidean distance between x and y
}
\description{
Euclidean distance
}
\examples{
euclidean_distance(c(1, 1), c(1, 2))
}
