% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_stopping_condition.R
\name{gsc_metaepochs_count}
\alias{gsc_metaepochs_count}
\title{Factory function for a global stopping condition that stops
the computation after given number of metaepochs.}
\usage{
gsc_metaepochs_count(metaepochs_count)
}
\arguments{
\item{metaepochs_count}{\itemize{
\item numeric - maximum number of metaepochs
}}
}
\value{
Function that receives a list of metaepoch snapshots
and returns a Boolean value determining whether the computation
should be stopped based on how many metaepochs have passed,
which can be used as a global stopping condition for the hms function.
}
\description{
Factory function for a global stopping condition that stops
the computation after given number of metaepochs.
}
\examples{
global_stopping_condition <- gsc_metaepochs_count(10)
}
