% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\docType{class}
\name{hms-class}
\alias{hms-class}
\title{A S4 class representing a result of hms.}
\description{
A S4 class representing a result of hms.
}
\section{Slots}{

\describe{
\item{\code{root_id}}{character - UUID of a root Deme.}

\item{\code{metaepoch_snapshots}}{list of objects of class MetaepochSnapshot.}

\item{\code{best_fitness}}{numeric - best fitness value of all metaepochs.}

\item{\code{best_solution}}{numeric -  best solution of all metaepochs.}

\item{\code{total_time_in_seconds}}{numeric - time of a hms execution in seconds.}

\item{\code{total_metaepoch_time_in_seconds}}{numeric - time of all metaepochs in seconds.}

\item{\code{metaepochs_count}}{numeric - total number of all metaepochs.}

\item{\code{deme_population_sizes}}{numeric - sizes of deme populations on each tree level.
Same as \code{population_sizes} parameter of hms function.}

\item{\code{lower}}{numeric - lower bound of the domain, a vector of length equal
to the decision variables.}

\item{\code{upper}}{numeric - upper bound of the domain, a vector of length equal
to the decision variables.}

\item{\code{call}}{language - an object of class "call" representing the matched call.}
}}

