% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\docType{package}
\name{hms-package}
\alias{hms-package}
\title{hms: Pretty Time of Day}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements an S3 class for storing and formatting time-of-day values, based on the 'difftime' class.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hms.tidyverse.org/}
  \item \url{https://github.com/tidyverse/hms}
  \item Report bugs at \url{https://github.com/tidyverse/hms/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

Other contributors:
\itemize{
  \item R Consortium [funder]
  \item RStudio [funder]
}

}
