% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\docType{package}
\name{hms-package}
\alias{hms-package}
\title{hms: Pretty Time of Day}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Implements an S3 class for storing and formatting time-of-day
    values, based on the 'difftime' class.
}
\details{
\Sexpr[results=rd, stage=render]{hms:::lifecycle("stable")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hms.tidyverse.org/}
  \item \url{https://github.com/tidyverse/hms}
  \item Report bugs at \url{https://github.com/tidyverse/hms/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Other contributors:
\itemize{
  \item The R Consortium [funder]
  \item RStudio [funder]
}

}
