% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\name{hms}
\alias{as.POSIXct.hms}
\alias{as.POSIXlt.hms}
\alias{as.character.hms}
\alias{as.hms}
\alias{as.hms.POSIXt}
\alias{as.hms.character}
\alias{as.hms.default}
\alias{as.hms.difftime}
\alias{as.hms.numeric}
\alias{format.hms}
\alias{hms}
\alias{is.hms}
\alias{print.hms}
\title{A simple class for storing time-of-day values}
\usage{
hms(seconds = 0, minutes = 0, hours = 0, days = 0)

is.hms(x)

as.hms(x, ...)

\method{as.hms}{default}(x, ...)

\method{as.hms}{difftime}(x, ...)

\method{as.hms}{numeric}(x, ...)

\method{as.hms}{character}(x, ...)

\method{as.hms}{POSIXt}(x, ...)

\method{as.POSIXct}{hms}(x, ...)

\method{as.POSIXlt}{hms}(x, ...)

\method{as.character}{hms}(x, ...)

\method{format}{hms}(x, ...)

\method{print}{hms}(x, ...)
}
\arguments{
\item{seconds, minutes, hours, days}{Time since midnight. No bounds checking is
performed.}

\item{x}{An object.}

\item{...}{Arguments passed on to further methods.}
}
\description{
The values are stored as a \code{\link{difftime}} vector with a custom class.
Supports construction from time values, coercion to and from
various data types, and formatting.  Can be used as a regular column in a
data frame.
}
\examples{
hms(56, 34, 12)
as.hms(1)
as.hms("12:34:56")
as.hms(Sys.time())
as.POSIXct(hms(1))
\dontrun{
  # Will raise an error
  data.frame(a = hms(1))
}
d <- data.frame(hours = 1:3)
d$hours <- hms(hours = d$hours)
d
}

