% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\name{hms}
\alias{hms}
\title{Title}
\usage{
hms(
  max_tree_height = 5,
  fitness,
  lower,
  upper,
  sigma,
  population_size,
  run_metaepoch = ga_metaepoch(list(list(), list(), list(), list(), list())),
  global_stopping_condition = default_global_stopping_condition,
  local_stopping_condition = default_local_stopping_condition,
  sprouting_condition = max_euclidean_distance_sprouting_condition(0.5)
)
}
\arguments{
\item{sprouting_condition}{}
}
\value{

}
\description{
Title
}
