% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\name{hms}
\alias{hms}
\title{Title}
\usage{
hms(
  tree_height = 5,
  fitness,
  lower,
  upper,
  sigma = default_sigma(lower, upper, tree_height),
  population_size_per_tree_level = rep(10, tree_height),
  run_metaepoch = ga_metaepoch(list(list(), list(), list(), list(), list())),
  global_stopping_condition = default_global_stopping_condition,
  local_stopping_condition = default_local_stopping_condition,
  sprouting_condtion_distance = euclidean_distance,
 
    sprouting_condition_distance_per_tree_level = sprouting_condition_default_euclidean_distances(lower,
    upper, sigma),
  sprouting_condition = max_metric_sprouting_condition(sprouting_condtion_distance,
    sprouting_condition_distance_per_tree_level),
  create_population,
  suggestions = NULL,
  with_gradient_method = FALSE,
  gradient_method_args = list(method = "L-BFGS-B", poptim = 0.05, pressel = 0.5,
    control = list(fnscale = -1, maxit = 100)),
  run_gradient_method,
  monitor = FALSE
)
}
\arguments{
\item{tree_height}{\itemize{
\item numeric - default value: 5
}}

\item{fitness}{\itemize{
\item fitness function
}}

\item{lower}{\itemize{
\item numeric - lower bound
}}

\item{upper}{\itemize{
\item numeric - upper bound
}}

\item{sigma}{\itemize{
\item numeric - vector of standard deviations for each tree level
}}

\item{run_metaepoch}{\itemize{
\item function - returns list with named fields: solution, population, value
}}

\item{global_stopping_condition}{\itemize{
\item function
}}

\item{local_stopping_condition}{\itemize{
\item function
}}

\item{sprouting_condition}{\itemize{
\item function
}}

\item{create_population}{\itemize{
\item function
}}

\item{suggestions}{\itemize{
\item matrix
}}

\item{with_gradient_method}{\itemize{
\item logical
}}

\item{gradient_method_args}{\itemize{
\item list of parameters that are passed to the gradient method
}}

\item{run_gradient_method}{\itemize{
\item function - returns list with named fields: solution, population, value
}}

\item{population_size}{\itemize{
\item numeric
}}
}
\value{
numeric best solution
}
\description{
Title
}
