% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\name{hms}
\alias{hms}
\title{Title}
\usage{
hms(
  tree_height = 5,
  fitness,
  lower,
  upper,
  sigma,
  population_size = 20,
  run_metaepoch = ga_metaepoch(list(list(), list(), list(), list(), list())),
  global_stopping_condition = default_global_stopping_condition,
  local_stopping_condition = default_local_stopping_condition,
  sprouting_condition = max_metric_sprouting_condition(euclidean_distance, 0.5),
  create_population,
  suggestions = NULL
)
}
\arguments{
\item{tree_height}{\itemize{
\item numeric - default value: 5
}}

\item{fitness}{\itemize{
\item fitness function
}}

\item{lower}{\itemize{
\item numeric - lower bound
}}

\item{upper}{\itemize{
\item numeric - upper bound
}}

\item{sigma}{\itemize{
\item numeric - vector of standard deviations for each tree level
}}

\item{population_size}{\itemize{
\item numeric
}}

\item{run_metaepoch}{\itemize{
\item function - returns list with named fields: solution, population, value
}}

\item{global_stopping_condition}{\itemize{
\item function
}}

\item{local_stopping_condition}{\itemize{
\item function
}}

\item{sprouting_condition}{\itemize{
\item function
}}

\item{create_population}{\itemize{
\item function
}}

\item{suggestions}{\itemize{
\item matrix
}}
}
\value{
numeric best solution
}
\description{
Title
}
