% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\name{hms}
\alias{hms}
\title{Title}
\usage{
hms(
  tree_height = 3,
  fitness,
  lower,
  upper,
  sigma = default_sigma(lower, upper, tree_height),
  population_size_per_tree_level = default_population_size_per_tree_level(tree_height),
  run_metaepoch = default_ga_metaepoch(tree_height),
  global_stopping_condition = default_global_stopping_condition,
  local_stopping_condition = default_local_stopping_condition,
  sprouting_condition = max_metric_sprouting_condition(euclidean_distance,
    sprouting_default_euclidean_distances(sigma)),
  create_population = default_create_population(sigma),
  suggestions = NULL,
  with_gradient_method = FALSE,
  gradient_method_args = default_gradient_method_args,
  run_gradient_method,
  monitor_level = "basic"
)
}
\arguments{
\item{tree_height}{\itemize{
\item numeric - default value: 5
}}

\item{fitness}{\itemize{
\item fitness function
}}

\item{lower}{\itemize{
\item numeric - lower bound
}}

\item{upper}{\itemize{
\item numeric - upper bound
}}

\item{sigma}{\itemize{
\item numeric - vector of standard deviations for each tree level
}}

\item{population_size_per_tree_level}{\itemize{
\item numeric
}}

\item{run_metaepoch}{\itemize{
\item function - returns list with named fields: solution, population, value
}}

\item{global_stopping_condition}{\itemize{
\item function
}}

\item{local_stopping_condition}{\itemize{
\item function
}}

\item{sprouting_condition}{\itemize{
\item function
}}

\item{create_population}{\itemize{
\item function
}}

\item{suggestions}{\itemize{
\item matrix
}}

\item{with_gradient_method}{\itemize{
\item logical
}}

\item{gradient_method_args}{\itemize{
\item list of parameters that are passed to the gradient method
}}

\item{run_gradient_method}{\itemize{
\item function - returns list with named fields: solution, population, value
}}

\item{monitor_level}{\itemize{
\item string - one of {'none', 'basic', 'basic_tree', 'verbose_tree'}
}}
}
\value{
numeric best solution
}
\description{
Title
}
