% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stopping_condition.R
\name{lsc_max_fitness_evaluations}
\alias{lsc_max_fitness_evaluations}
\title{Factory function for a local stopping condition that stops a deme
after given number of fitness function evaluations has been made
in that deme.}
\usage{
lsc_max_fitness_evaluations(max_evaluations)
}
\arguments{
\item{max_evaluations}{\itemize{
\item numeric
}}
}
\value{
Function that can be used as a local stopping condition for hms.
}
\description{
Factory function for a local stopping condition that stops a deme
after given number of fitness function evaluations has been made
in that deme.
}
\examples{
local_stopping_condition <- lsc_max_fitness_evaluations(500)
}
