% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stopping_condition.R
\name{lsc_metaepochs_without_improvement}
\alias{lsc_metaepochs_without_improvement}
\title{Factory function for a local stopping condition that stops a deme
after given number of consecutive metaeopochs without an improvement
of the best solution found in that deme.}
\usage{
lsc_metaepochs_without_improvement(max_metaepochs_without_improvement)
}
\arguments{
\item{max_metaepochs_without_improvement}{\itemize{
\item numeric
}}
}
\value{
Function that can be used as a local stopping condition for hms.
}
\description{
Factory function for a local stopping condition that stops a deme
after given number of consecutive metaeopochs without an improvement
of the best solution found in that deme.
}
\examples{
local_stopping_condition <- lsc_metaepochs_without_improvement(5)
}
