% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{manhattan_distance}
\alias{manhattan_distance}
\title{Manhattan distance}
\usage{
manhattan_distance(x, y)
}
\arguments{
\item{x}{\itemize{
\item numeric
}}

\item{y}{\itemize{
\item numeric
}}
}
\value{
numeric - manhattan distance between x and y
}
\description{
Manhattan distance
}
\examples{
manhattan_distance(c(1, 1), c(1, 2))
}
