% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\docType{class}
\name{MetaepochSnapshot-class}
\alias{MetaepochSnapshot-class}
\title{A S4 class representing a snapshot of one metaepoch.}
\description{
A S4 class representing a snapshot of one metaepoch.
}
\section{Slots}{

\describe{
\item{\code{demes}}{list of objects of class Deme.}

\item{\code{best_fitness}}{numeric - best fitness value of a metaepoch.}

\item{\code{best_solution}}{numeric - best solution of a metaepoch.}

\item{\code{time_in_seconds}}{numeric - time of metaepoch in seconds.}

\item{\code{fitness_evaluations}}{numeric - number of fitness evaluations.}

\item{\code{blocked_sprouts}}{list - list of sprouts that were blocked by sprouting
condition. A sprout is a potential origin of a new Deme, it can be blocked
by sc -- sprouting condition. See \code{\link{sc_max_metric}} for more details.}

\item{\code{is_evolutionary}}{logical - \code{TRUE} for all metaepochs except the
gradient one.}
}}

