% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_hms}
\alias{parse_hms}
\title{Parsing hms values}
\usage{
parse_hms(x)
}
\arguments{
\item{x}{A character vector}
}
\description{
This function convertd character vectors to objects of the \link{hms} class.
\code{NA} values are supported.
}
\details{
\code{parse_hms()} accepts values of the form \code{"DdHH:MM:SS"}, with optional
days, and fractional seconds. It also supports minus signs.
}
\examples{
time_str <- c("12:34:56",         # regular format
              "12:34:56.789",     # decimal points
              "20:12",            # HH:MM (no seconds)
              "-12:34:56.001",    # negative values without days
              "1d 12:34:56",      # number of days (integer)
              "1d 12:34:56.001",  # decimal points and days
              "-1d 12:34:56.001", # negative values
              "-1d 12:34:56.001", # negative values and space
              "0.5d"              # day only
)

parse_hms(time_str)
}
