% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{plotPopulation}
\alias{plotPopulation}
\title{plotPopulation method for "hms" class.}
\usage{
plotPopulation(object, dimensions)
}
\arguments{
\item{object}{\itemize{
\item hms s4 object
}}

\item{dimensions}{\itemize{
\item two selected dimensions
}}
}
\value{
It doesn't return anything meaningful. It plots the selected two
dimensions of a population.
}
\description{
plotPopulation method for "hms" class.
}
\examples{
f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
plotPopulation(result, c(1, 1))
}
