% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{printBlockedSprouts}
\alias{printBlockedSprouts}
\title{printBlockedSprouts method for "hms" class.}
\usage{
printBlockedSprouts(object)
}
\arguments{
\item{object}{\itemize{
\item hms s4 object
}}
}
\value{
It doesn't return anything. It prints blocked sprouts per metaepoch.
}
\description{
printBlockedSprouts method for "hms" class.
}
\examples{
f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
printBlockedSprouts(result)
}
