% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round_hms}
\alias{round_hms}
\alias{trunc_hms}
\title{Round or truncate to a multiple of seconds}
\usage{
round_hms(x, secs, digits = NULL)

trunc_hms(x, secs, digits = NULL)
}
\arguments{
\item{x}{A vector of class \link{hms}}

\item{secs}{Multiple of seconds, a positive numeric. Values less than one
are supported}
}
\value{
The input, rounded or truncated to the nearest multiple of \code{secs}
}
\description{
Convenience functions to round or truncate to a multiple of seconds.
}
\examples{
round_hms(as_hms("12:34:56"), 5)
round_hms(as_hms("12:34:56"), 60)
round_hms(as_hms("12:34:56"), digits = 1)
round_hms(as_hms("12:34:56"), digits = -2)
trunc_hms(as_hms("12:34:56"), 60)
}
