% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{rtnorm_mutation}
\alias{rtnorm_mutation}
\title{Title}
\usage{
rtnorm_mutation(lower, upper, sd)
}
\arguments{
\item{lower}{\itemize{
\item numeric
}}

\item{upper}{\itemize{
\item numeric
}}

\item{sd}{\itemize{
\item numeric
}}
}
\value{

}
\description{
Title
}
