% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{saveMetaepochsPopulations}
\alias{saveMetaepochsPopulations}
\title{Title}
\usage{
saveMetaepochsPopulations(object, path, dimensions)
}
\arguments{
\item{object}{hms s4 object}

\item{path}{path}

\item{dimensions}{vector of two selected dimensions e.g. c(1,2)}
}
\description{
Title
}
