% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_class.R
\name{summary,hms-method}
\alias{summary,hms-method}
\title{Summary method for class "hms".}
\usage{
\S4method{summary}{hms}(object, ...)
}
\arguments{
\item{object}{\itemize{
\item hms s4 object
}}

\item{...}{\itemize{
\item other summary arguments
}}
}
\value{
Returns a list with fields: fitness, solution, metaepochs, deme_population_sizes,
lower_bound, upper_bound, computation_time. These fields should match fields
of class "hms".
}
\description{
Summary method for class "hms".
}
\examples{
f <- function(x) x
result <- hms(fitness = f, lower = -5, upper = 5)
summary(result)
}
